[Style] nested enum Constants를 하나의 파일로 작성하는 것으로 수정
