--- travelPlan/Source/Feature/Feed/Constant/PostSearch/CollectionReusableView/PostRecommendationSearchHeaderView+Constants.swift
@@ -1,16 +1,0 @@
-//
-//  PostRecommendationSearchHeaderView+Constants.swift
-//  travelPlan
-//
-//  Created by SeokHyun on 2023/08/31.
-//
-
-import Foundation
-
-extension PostRecommendationSearchHeaderView {
-  enum Constants {
-    enum TitleLabel {
-      static let fontSize: CGFloat = 17
-    }
-  }
-}
