--- travelPlan/Source/Feature/Feed/Constant/PostSearch/CollectionReusableView/PostSearchFooterView+Constants.swift
@@ -1,19 +1,0 @@
-//
-//  PostSearchFooterView+Constants.swift
-//  travelPlan
-//
-//  Created by SeokHyun on 2023/08/09.
-//
-
-import Foundation
-
-extension PostSearchFooterView {
-  enum Constants {
-    enum Inset {
-      static let leading: CGFloat = 0
-      static let top: CGFloat = 10
-      static let trailing: CGFloat = 0
-      static let bottom: CGFloat = 10
-    }
-  }
-}
