--- travelPlan/Source/Feature/Feed/Constant/PostSearch/PostSearchTagLabel+Constants.swift
@@ -1,15 +1,0 @@
-//
-//  PostSearchTagLabel+Constants.swift
-//  travelPlan
-//
-//  Created by SeokHyun on 2023/09/01.
-//
-
-import Foundation
-
-extension PostSearchTagLabel {
-  enum Constants {
-    static let fontSize: CGFloat = 14
-    static let numberOfLines = 1
-  }
-}
