--- travelPlan/Source/Feature/Feed/Constant/PostSearch/PostSearchViewModel+Constants.swift
@@ -1,18 +1,0 @@
-//
-//  PostSearchViewModel+Constants.swift
-//  travelPlan
-//
-//  Created by SeokHyun on 2023/08/10.
-//
-
-import Foundation
-
-extension PostSearchViewModel {
-  enum Constants {
-    enum CollectionView {
-      static let edgeInsetWidth: CGFloat = PostSearchViewController.Constants
-        .CollectionViewLayout.Inset.left + PostSearchViewController.Constants
-        .CollectionViewLayout.Inset.right
-    }
-  }
-}
