--- travelPlan/Source/Feature/Feed/Constant/PostSearch/TagCell/PostRecommendationSearchTagCell+Constants.swift
@@ -1,27 +1,0 @@
-//
-//  PostRecommendationSearchTagCell+Constants.swift
-//  travelPlan
-//
-//  Created by SeokHyun on 2023/08/10.
-//
-
-import Foundation
-
-extension PostRecommendationSearchTagCell {
-  enum Constants {
-    enum ContentView {
-      static let borderWidth: CGFloat = 1
-      static let cornerRadius: CGFloat = 15
-      static let alphaComponent: CGFloat = 0.1
-    }
-    
-    enum TagLabel {
-      enum Inset {
-        static let trailing: CGFloat = 13
-        static let leading: CGFloat = 13
-        static let top: CGFloat = 4
-        static let bottom: CGFloat = 4
-      }
-    }
-  }
-}
