--- travelPlan/Source/Feature/Search/Constants/SearchStarButton+Constants.swift
@@ -1,15 +1,0 @@
-//
-//  SearchStarButton+Constants.swift
-//  travelPlan
-//
-//  Created by SeokHyun on 2023/08/17.
-//
-
-import Foundation
-
-extension SearchStarButton {
-  enum Constants {
-    static let normalImageName = "emptyStar"
-    static let selectedImageName = "star"
-  }
-}
