+++ travelPlan/Source/Feature/Search/View/SearchFamousSpotCell.swift
@@ -10,6 +10,48 @@ import SnapKit
 import Combine
 
 final class SearchFamousSpotCell: UICollectionViewCell {
+  enum Constants {
+    // MARK: - ThumbnailImageView
+    enum ThumbnailImageView {
+      static let imageName = "tempThumbnail1"
+      static let cornerRadius: CGFloat = 3
+      enum Inset {
+        static let top: CGFloat = 5
+        static let bottom: CGFloat = 5
+      }
+    }
+    // MARK: - LabelStackView
+    enum LabelStackView {
+      static let spacing: CGFloat = 0
+      enum Offset {
+        static let leading: CGFloat = 15
+        static let trailing: CGFloat = -20
+      }
+    }
+    // MARK: - HeartButton
+    enum HeartButton {
+      static let size: CGFloat = 24
+      enum Inset {
+        static let top: CGFloat = 5
+      }
+    }
+    // MARK: - PlaceLabel
+    enum PlaceLabel {
+      static let fontSize: CGFloat = 16
+      static let numberOfLines = 1
+    }
+    // MARK: - CategoryLabel
+    enum CategoryLabel {
+      static let fontSize: CGFloat = 14
+      static let numberOfLines = 1
+    }
+    // MARK: - AreaLabel
+    enum AreaLabel {
+      static let size: CGFloat = 14
+      static let numberOfLines = 1
+    }
+  }
+  
   // MARK: - Properties
   private var viewModel: SearchFamousSpotCellViewModel? {
     didSet {
@@ -28,7 +70,7 @@ final class SearchFamousSpotCell: UICollectionViewCell {
     $0.layer.masksToBounds = true
   }
   
-  private lazy var starButton: SearchStarButton = .init().set {
+  private lazy var starButton: SearchStarButton = .init(normalType: .empty).set {
     $0.addTarget(self, action: #selector(didTapStarButton), for: .touchUpInside)
   }
