+++ travelPlan/Source/Feature/Search/ViewController/SearchViewController.swift
@@ -11,6 +11,26 @@ import UIKit
 import SnapKit
 
 final class SearchViewController: UIViewController {
+  enum Constants {
+    // MARK: - SearchView
+    enum SearchView {
+      enum Spacing {
+        static let top: CGFloat = 20
+        static let leading: CGFloat = 16
+        static let trailing: CGFloat = 16
+      }
+      static let height: CGFloat = 50
+    }
+    // MARK: - CollectionView
+    enum CollectionView {
+      enum Spacing {
+        enum Offset {
+          static let top: CGFloat = 20
+        }
+      }
+    }
+  }
+  
   // MARK: - Properties
   weak var coordinator: SearchCoordinator?
   private let viewModel = SearchViewModel()
